Bug fixing and moving string helpers to common class.
