+++ qrsInteractFactory.js
@@ -7,25 +7,7 @@ var qrsInteractMain = require('./qrsInteract');
 
 
 var qrsInteract = function QRSInteract(inputConfig) {
-
-    if (!String.prototype.startsWith) {
-        String.prototype.startsWith = function (searchString, position) {
-            position = position || 0;
-            return this.substr(position, searchString.length) === searchString;
-        };
-    }
-
-    if (!String.prototype.endsWith) {
-        String.prototype.endsWith = function (searchString, position) {
-            var subjectString = this.toString();
-            if (typeof position !== 'number' || !isFinite(position) || Math.floor(position) !== position || position > subjectString.length) {
-                position = subjectString.length;
-            }
-            position -= searchString.length;
-            var lastIndex = subjectString.indexOf(searchString, position);
-            return lastIndex !== -1 && lastIndex === position;
-        };
-    }
+    common.initStringHelpers();
 
     var updateConfig = function (inputConfig) {
         var newConfig = common.clone(config);
@@ -44,12 +26,14 @@ var qrsInteract = function QRSInteract(inputConfig) {
             });
         }
 
-        if (newConfig.virtualProxyPrefix != "" && !newConfig.virtualProxyPrefix.startsWith('/')) {
-            newConfig.virtualProxyPrefix = '/' + newConfig.virtualProxyPrefix;
-        }
+        if (newConfig.virtualProxyPrefix != undefined) {
+            if (newConfig.virtualProxyPrefix != "" && !newConfig.virtualProxyPrefix.startsWith('/')) {
+                newConfig.virtualProxyPrefix = '/' + newConfig.virtualProxyPrefix;
+            }
 
-        if (newConfig.virtualProxyPrefix != "" && newConfig.virtualProxyPrefix.endsWith('/')) {
-            newConfig.virtualProxyPrefix = newConfig.virtualProxyPrefix.substr(0, newConfig.virtualProxyPrefix.length - 1);
+            if (newConfig.virtualProxyPrefix != "" && newConfig.virtualProxyPrefix.endsWith('/')) {
+                newConfig.virtualProxyPrefix = newConfig.virtualProxyPrefix.substr(0, newConfig.virtualProxyPrefix.length - 1);
+            }
         }
 
         return newConfig;
